1.0: separate code and docs
