+++ Makefile
@@ -1,17 +1,18 @@
+PROJECT=vinyldc
 CC=gcc
-CFLAGS=-c -Wall -lSDL2  -lSDL2_image -lm -g
-LDFLAGS=-lSDL2  -lSDL2_image -lm -g
-SOURCES=main.c wav.c vinyl.c point.c surface.c
+CFLAGS=-c -Wall -Isrc -g
+LDFLAGS=-lSDL2 -lSDL2_image -lm
+SOURCES=src/main.c src/wav.c src/vinyl.c src/point.c src/surface.c
 OBJECTS=$(SOURCES:.c=.o)
-EXECUTABLE=vinyldc
+RM=rm -f
 
-all: $(SOURCES) $(EXECUTABLE)
-	
-$(EXECUTABLE): $(OBJECTS) 
-	$(CC)  $(OBJECTS) $(LDFLAGS) -o $@
+all: $(PROJECT)
 
-.c.o:
+$(PROJECT): $(OBJECTS) 
+	$(CC) $^ $(LDFLAGS) -o $@
+
+%.o: %.c
 	$(CC) $(CFLAGS) $< -o $@
 
 clean:
-	rm -f $(OBJECTS) $(EXECUTABLE)
+	$(RM) $(OBJECTS) $(PROJECT)
