+++ src/vinyl.h
