--- vinyl.c
@@ -1,35 +1,0 @@
-#include <SDL2/SDL.h>
-#include <SDL2/SDL_image.h>
-
-#include "vinyl.h"
-
-bool vinyl_init()
-{
-	bool success = true;
-
-	if (SDL_Init(SDL_INIT_VIDEO) < 0) {
-		printf( "Failed to initialize SDL: %s\n", SDL_GetError());
-		success = false;
-		goto out;
-	}
-
-	int imgFlags = IMG_INIT_PNG;
-	if (!(IMG_Init(imgFlags) & imgFlags)) {
-		printf( "Failed to initialize SDL_image: %s\n", IMG_GetError() );
-		success = false;
-		goto img_init_fail;
-	}
-
-out:
-	return success;
-img_init_fail:
-	SDL_Quit();
-	return success;
-}
-
-
-void vinyl_exit()
-{
-	IMG_Quit();
-	SDL_Quit();
-}
